Move mock component out of camel-core. Work in progress.
