+++ core/camel-core/pom.xml
@@ -107,6 +107,10 @@
         </dependency>
         <dependency>
             <groupId>org.apache.camel</groupId>
+            <artifactId>camel-dataset</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>org.apache.camel</groupId>
             <artifactId>camel-direct</artifactId>
         </dependency>
         <dependency>
