+++ core/camel-core/src/test/data/file-dataset-test.txt
