+++ core/camel-core/src/test/java/org/apache/camel/component/dataset/DataSetTestAnyOrderTest.java
@@ -16,43 +16,33 @@
  */
 package org.apache.camel.component.dataset;
 
+import org.apache.camel.ContextTestSupport;
 import org.apache.camel.builder.RouteBuilder;
-import org.apache.camel.test.junit4.CamelTestSupport;
-import org.junit.Before;
-import org.junit.Ignore;
 import org.junit.Test;
 
-public class DataSetTestFileSplitTest extends CamelTestSupport {
+public class DataSetTestAnyOrderTest extends ContextTestSupport {
 
     @Override
     public boolean isUseRouteBuilder() {
         return false;
     }
 
-    @Override
-    @Before
-    public void setUp() throws Exception {
-        deleteDirectory("target/data/testme");
-        super.setUp();
-    }
-
-    @Ignore
     @Test
-    public void testFile() throws Exception {
-        template.sendBody("file:target/data/testme", "Hello World\nBye World\nHi World");
+    public void testAnyOrder() throws Exception {
+        template.sendBody("seda:testme", "Bye World");
+        template.sendBody("seda:testme", "Hello World");
 
         context.addRoutes(new RouteBuilder() {
             @Override
             public void configure() throws Exception {
                 from("direct:start")
-                        .to("dataset-test:file:target/data/testme?noop=true&split=true&timeout=1000");
+                        .to("dataset-test:seda:testme?anyOrder=true&timeout=0");
             }
         });
         context.start();
 
         template.sendBody("direct:start", "Hello World");
         template.sendBody("direct:start", "Bye World");
-        template.sendBody("direct:start", "Hi World");
 
         assertMockEndpointsSatisfied();
     }
