+++ core/camel-core/src/test/java/org/apache/camel/component/dataset/DataSetTestFileTest.java
@@ -16,31 +16,43 @@
  */
 package org.apache.camel.component.dataset;
 
-import org.apache.camel.BindToRegistry;
+import org.apache.camel.ContextTestSupport;
 import org.apache.camel.builder.RouteBuilder;
-import org.apache.camel.test.junit4.CamelTestSupport;
+import org.junit.Before;
+import org.junit.Ignore;
 import org.junit.Test;
 
-public class DataSetTest extends CamelTestSupport {
+public class DataSetTestFileTest extends ContextTestSupport {
 
-    @BindToRegistry("foo")
-    protected SimpleDataSet dataSet = new SimpleDataSet(20);
-
-    @Test
-    public void testDataSet() throws Exception {
-        // data set will itself set its assertions so we should just
-        // assert that all mocks is ok
-        assertMockEndpointsSatisfied();
+    @Override
+    public boolean isUseRouteBuilder() {
+        return false;
     }
 
     @Override
-    protected RouteBuilder createRouteBuilder() throws Exception {
-        return new RouteBuilder() {
-            public void configure() throws Exception {
-                // start this first to make sure the "direct:foo" consumer is ready
-                from("direct:foo").to("dataset:foo?minRate=50");
-                from("dataset:foo?initialDelay=0&minRate=50").to("direct:foo");
-            }
-        };
+    @Before
+    public void setUp() throws Exception {
+        deleteDirectory("target/data/testme");
+        super.setUp();
     }
+
+    @Ignore
+    @Test
+    public void testFile() throws Exception {
+        template.sendBody("file:target/data/testme", "Hello World");
+
+        context.addRoutes(new RouteBuilder() {
+            @Override
+            public void configure() throws Exception {
+                from("direct:start")
+                        .to("dataset-test:file:target/data/testme?noop=true&timeout=1500");
+            }
+        });
+        context.start();
+
+        template.sendBody("direct:start", "Hello World");
+
+        assertMockEndpointsSatisfied();
+    }
+
 }
